Refactored Image Resizer.
